% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{correlation_scatter}
\alias{correlation_scatter}
\title{Plot a Scatter Plot with Correlation Coefficient}
\usage{
correlation_scatter(
  data,
  x = NULL,
  y = NULL,
  group = NULL,
  cor_method = "pearson",
  use = "everything",
  verbose = TRUE
)
}
\arguments{
\item{data}{A tibble or dataframe}

\item{x}{First numeric variable}

\item{y}{Second numeric variable}

\item{group}{Variable used for grouping}

\item{cor_method}{Which correlation coefficient to calculate}

\item{use}{Method for handling missing values}

\item{verbose}{Print relevant information}
}
\value{
a ggplot object
}
\description{
Plot a scatter plot with correlation coefficient added as annotations.
Wrapper around the \code{ggplot} and \code{cor} function.
}
\examples{
correlation_scatter(iris, x="Sepal.Length", y="Sepal.Width", group="Species")
correlation_scatter(iris, cor_method="pearson")
}
