\name{corrplot-package}
\Rdversion{1.1}
\alias{corrplot-package}
\alias{corrplot}
\docType{package}
\title{
visualization of a correlation matrix 
}
\description{
The corrplot package is a graphical display of a correlation matrix, test for correlation, and other visualization methods about association and correlation.
}
\details{
\tabular{ll}{
Package: \tab corrplot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-06-22\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Taiyun Wei

Maintainer:Taiyun Wei <weitaiyun@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
The \code{plotcorr} function in the \code{ellipse} package and \code{corrgram} function in the \code{corrgram} package has some similarities.
}
