% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.pair.R
\name{corrplot.pair}
\alias{corrplot.pair}
\title{Using mixed methods to visualize a correlation matrix.}
\usage{
corrplot.pair(corr1, corr2, lower = "circle", upper = "circle",
  col1 = NULL, col2 = NULL, order = c("original", "AOE", "FPC", "hclust",
  "alphabet"), order.guide = c("l", "u"), hclust.method = c("complete",
  "ward", "ward.D", "ward.D2", "single", "average", "mcquitty", "median",
  "centroid"), tl.pos = c("d", "lt", "n"), tl.cex = 1, cl.offset = c(2,
  0.5), diag = c("n", "l", "u"), bg = "white", addgrid.col = "grey",
  p.mat1 = NULL, sig.level1 = 0.05, p.mat2 = NULL, sig.level2 = 0.05,
  sig = c("pch", "p-value", "blank", "n"), insig = c("pch", "p-value",
  "blank", "n"), pch1 = 4, pch1.col = "black", pch1.cex = 3, pch2 = 4,
  pch2.col = "black", pch2.cex = 3, plotCI = c("n", "square", "circle",
  "rect"), ...)
}
\arguments{
\item{lower}{Character, the visualization method for the lower triangular
correlation matrix.}

\item{upper}{Character, the visualization method for the upper triangular
correlation matrix.}

\item{tl.pos}{Character, \code{"lt"}, \code{"d"} or \code{"n"}, giving
position of text labels, \code{"lt"} means left and top,  \code{"d"} means
diagonal. If \code{"n"},  add no textlabel.}

\item{diag}{Character, for specifying the glyph on the principal diagonal. It
is one of \code{"n"} (default,  draw nothing), \code{"l"} (draw the glyphs
of lower triangular) or \code{"u"} (draw the glyphs of upper triangular).}

\item{bg}{The background color.}

\item{addgrid.col}{See the \code{addgrid.col} parameter in the function
\code{\link{corrplot}}}

\item{plotCI}{See the \code{plotCI} parameter in the function
\code{\link{corrplot}}}

\item{\dots}{Additional arguments for corrplot's wrappers}

\item{corr}{Matrix, the correlation matrix to visualize.}
}
\description{
Using mixed methods to visualize a correlation matrix.
}
\examples{
M <- cor(mtcars)
ord <- corrMatOrder(M, order = "AOE")
M2 <- M[ord,ord]


corrplot.mixed(M2)
corrplot.mixed(M2, lower = "ellipse", upper = "circle")
corrplot.mixed(M2, lower = "square", upper = "circle")
corrplot.mixed(M2, lower = "shade", upper = "circle")
corrplot.mixed(M2, tl.pos = "lt")
corrplot.mixed(M2, tl.pos = "lt", diag = "u")
corrplot.mixed(M2, tl.pos = "lt", diag = "l")
corrplot.mixed(M2, tl.pos = "n")
}
\author{
Taiyun Wei
}
