% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplotVis.R
\name{corrplotVis}
\alias{corrplotVis}
\title{This is an plot function for visualizing a correlation matrix by using html5}
\usage{
corrplotVis(corr, color = NULL, size = c(950, 900), widthRange = c(100,
  1000))
}
\arguments{
\item{corr}{The correlation matrix to visualize.}

\item{color}{Vector, 3 color hexadecimal code the corresponding correlation value 1,0,-1.
When it is NULL, color will be \code{color = c("#053061", "#FFFFFF", "#67001F")}}

\item{size}{Vector, describe the width and the height of the visualiztion area.}

\item{widthRange}{Vector, decscribe the min and the max value of the area range.}
}
\description{
The function generates 'corrplot' html files viewed by Chrome which help you
using in a interative and convenient way.
}
\examples{
data(mtcars)
M <- cor(mtcars)

corrplot::corrplotVis(M)
}
