% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrRect.R
\name{corrRect}
\alias{corrRect}
\title{Draw rectangle(s) on the correlation matrix graph.}
\usage{
corrRect(clus = NULL, namesMat = NULL, corrRes, col = "black", lwd = 2, ...)
}
\arguments{
\item{clus}{Vector, the number of each cluster's members.}

\item{namesMat}{4-length character vector or 4-columns character matrix,
represents the names of xleft, ybottom, xright, ytop correspondingly.}

\item{corrRes}{list of the \code{corrplot()} returns.}

\item{col}{Color of rectangles.}

\item{lwd}{Line width of rectangles.}

\item{\dots}{Additional arguments passing to function \code{rect()}.}
}
\description{
Draw rectangle(s) around the chart of corrrlation matrix.
}
\details{
\code{corrRect} needs the number(parameter \code{clus}) of each cluster's
members, while \code{corrRect.hclust}  can get the members in each cluster
based on hierarchical clustering (\code{\link{hclust}}).
}
\examples{
data(mtcars)
M = cor(mtcars)


f = corrplot(M, order = "AOE")
r = rbind(c('gear', 'wt', 'qsec', 'carb'),
        c('wt', 'gear', 'carb', 'qsec'))
corrRect(namesMat=r, corrRes=f)


f = corrplot(M, method = "circle", order = "AOE", type='lower')
r = c('gear', 'carb', 'qsec', 'wt')
corrRect(namesMat=r, corrRes=f)


corrplot(M, method = "circle", order = "hclust")
corrRect(clus = c(5,6))

corrplot(M, method = "circle", order = "hclust")
corrRect(clus = c(5,3,3))

(order.hc = corrMatOrder(M, order = "hclust"))
(order.hc2 = corrMatOrder(M, order = "hclust", hclust.method = "ward.D"))
M.hc  = M[order.hc, order.hc]
M.hc2 = M[order.hc2, order.hc2]

par(ask = TRUE)

# same as: corrplot(M, order = "hclust", addrect = 2)
corrplot(M.hc)
corrRect.hclust(corr = M.hc, k = 2)

# same as: corrplot(M, order = "hclust", addrect = 3)
corrplot(M.hc)
corrRect.hclust(corr = M.hc, k = 3)

# same as: corrplot(M, order = "hclust", hclust.method = "ward.D", addrect = 2)
corrplot(M.hc2)
corrRect.hclust(M.hc2, k = 2, method = "ward.D")

# same as: corrplot(M, order = "hclust", hclust.method = "ward.D", addrect = 3)
corrplot(M.hc2)
corrRect.hclust(M.hc2, k = 3, method = "ward.D")

# same as: corrplot(M, order = "hclust", hclust.method = "ward.D", addrect = 4)
corrplot(M.hc2)
corrRect.hclust(M.hc2, k = 4, method = "ward.D")
}
\author{
Taiyun Wei
}
\keyword{hplot}
