% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot-rectangles.R
\name{make_rect}
\alias{make_rect}
\title{Add a rectangle to a correlation plot}
\usage{
make_rect(
  rstart,
  rend,
  cstart = NULL,
  cend = NULL,
  correlation_matrix,
  mirror = FALSE,
  lwd = 3,
  ...
)
}
\arguments{
\item{rstart}{The name of the variable contained in the correlation plot that will serve as the vertical starting point of the rectangle.}

\item{rend}{The name of the variable contained in the correlation plot that will serve as the vertical ending point of the rectangle.}

\item{cstart}{(optional) The name of the variable contained in the correlation plot that will serve as the horizontal starting point of the rectangle.
If no cstart or cend provided, creates an on-diagonal rectangle automatically. cstart and cend allows for off-diagonal rectangles.}

\item{cend}{(optional) The name of the variable contained in the correlation plot that will serve as the horizontal ending point of the rectangle.
If no cstart or cend provided, creates an on-diagonal rectangle automatically. cstart and cend allows for off-diagonal rectangles.}

\item{correlation_matrix}{The correlation matrix the correlation plot is based on.}

\item{mirror}{If TRUE, also adds the equivalent rectangle from the other side of the diagonal (i.e. the "mirror" of the original).
The function ignores this entirely if the rectangle is on the diagonal, as there is no mirror.}

\item{lwd}{Determines the width of the rectangle lines.}

\item{...}{Arguments to modify graphical parameters, etc.}
}
\description{
Add a rectangle to a correlation plot
}
\examples{
#Adding an on-diagonal rectangle to a correlation plot with mirroring, using
#the Bechtoldt sample correlation matrix provided by the psych package.
library(corrplot)
library(psych)
corrplot(Bechtoldt)
make_rect(rstart="First_Names", rend="Flags", correlation_matrix=Bechtoldt, mirror=TRUE)

#Adding an off-diagonal rectangle
make_rect(rstart="First_Names", rend="Flags", cstart="First_Names",
cend="Sentences", correlation_matrix=Bechtoldt)

}
