\name{corrplot-package}
\Rdversion{1.1}
\alias{corrplot-package}
\docType{package}
\title{
Visualization of a correlation matrix 
}
\description{
The corrplot package is a graphical display of a correlation matrix, 
confidence interval. It also contains some algorithms to do matrix reordering.
}
\details{
\tabular{ll}{
Package: \tab corrplot\cr
Type: \tab Package\cr
Version: \tab 0.66 \cr
Date: \tab 2012-08-14\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Taiyun Wei

Maintainer:Taiyun Wei <weitaiyun@gmail.com>
}
\references{
Michael Friendly (2002). 
\emph{Corrgrams: Exploratory displays for correlation matrices}.
The American Statistician, 56, 316--324.


D.J. Murdoch, E.D. Chow (1996).
\emph{A graphical display of large correlation matrices}.
The American Statistician, 50, 178--180.
}

\keyword{hplot}
\seealso{
The \code{plotcorr} function in the \code{ellipse} package and \code{corrgram} function in the \code{corrgram} package has some similarities.
}
