% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot-package.R
\docType{package}
\name{corrplot-package}
\alias{corrplot-package}
\alias{corrplot}
\title{Visualization of a correlation matrix}
\description{
The corrplot package is a graphical display of a correlation matrix,
confidence interval. It also contains some algorithms to do matrix
reordering.
}
\references{
Michael Friendly (2002).
\emph{Corrgrams: Exploratory displays for correlation matrices}.
The American Statistician, 56, 316--324.

D.J. Murdoch, E.D. Chow (1996).
\emph{A graphical display of large correlation matrices}.
The American Statistician, 50, 178--180.
}
\seealso{
The \code{plotcorr} function in the \code{ellipse} package and
\code{corrgram} function in the \code{corrgram} package has some
similarities.
}
\author{
Taiyun Wei (weitaiyun@gmail.com)

Viliam Simko (viliam.simko@gmail.com)

Maintainer: Taiyun Wei (weitaiyun@gmail.com)
}
\keyword{correlation}
\keyword{hplot}
